requirements.txt
httpx>=0.27
websockets>=12.0
anyio>=4.0
pydantic>=2.0
python-dotenv>=1.0
loguru>=0.7
PySide6>=6.6
PyYAML>=6.0
pytest>=8.0
